Updates code after review and adds decorators.py
