+++ usecase/test/test_download_dataset_as_zip_for_cron_job.py
@@ -2,23 +2,12 @@ from unittest import TestCase, mock
 from unittest.mock import MagicMock
 import datetime
 import os
-import warnings
 from representation.dataset_infos import DatasetInfos
 from usecase.download_dataset_as_zip import (
     download_dataset_as_zip_for_cron_job,
     add_download_date_for_cron_job,
 )
-
-
-def ignore_resource_warnings(test_func):
-    """Removes the resource warnings raised by testing download execution (normal class behaviour)."""
-
-    def test(self, *args, **kwargs):
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", ResourceWarning)
-            test_func(self, *args, **kwargs)
-
-    return test
+from utilities.decorators import ignore_resource_warnings
 
 
 class TestAddDownloadDateForCronJob(TestCase):
